+++ tests/BaGet.Tests/ApiIntegrationTests.cs
@@ -1,26 +1,26 @@
+using System;
 using System.IO;
 using System.Net;
 using System.Net.Http;
 using System.Threading.Tasks;
-using Microsoft.AspNetCore.Mvc.Testing;
 using Newtonsoft.Json;
 using Xunit;
 using Xunit.Abstractions;
 
 namespace BaGet.Tests
 {
-    public class ApiIntegrationTests : IClassFixture<BaGetWebApplicationFactory>
+    public class ApiIntegrationTests : IDisposable
     {
-        private readonly WebApplicationFactory<Startup> _factory;
+        private readonly BaGetApplication _app;
         private readonly HttpClient _client;
 
         private readonly Stream _packageStream;
         private readonly Stream _symbolPackageStream;
 
-        public ApiIntegrationTests(BaGetWebApplicationFactory factory, ITestOutputHelper output)
+        public ApiIntegrationTests(ITestOutputHelper output)
         {
-            _factory = factory.WithOutput(output);
-            _client = _factory.CreateClient();
+            _app = new BaGetApplication(output);
+            _client = _app.CreateClient();
 
             _packageStream = TestResources.GetResourceStream(TestResources.Package);
             _symbolPackageStream = TestResources.GetResourceStream(TestResources.SymbolPackage);
@@ -39,7 +39,7 @@ namespace BaGet.Tests
         [Fact]
         public async Task SearchReturnsOk()
         {
-            await _factory.AddPackageAsync(_packageStream);
+            await _app.AddPackageAsync(_packageStream);
 
             using var response = await _client.GetAsync("v3/search");
             var content = await response.Content.ReadAsStreamAsync();
@@ -101,7 +101,7 @@ namespace BaGet.Tests
         [Fact]
         public async Task AutocompleteReturnsOk()
         {
-            await _factory.AddPackageAsync(_packageStream);
+            await _app.AddPackageAsync(_packageStream);
 
             using var response = await _client.GetAsync("v3/autocomplete");
             var content = await response.Content.ReadAsStreamAsync();
@@ -139,7 +139,7 @@ namespace BaGet.Tests
         [Fact]
         public async Task AutocompleteVersionsReturnsOk()
         {
-            await _factory.AddPackageAsync(_packageStream);
+            await _app.AddPackageAsync(_packageStream);
 
             using var response = await _client.GetAsync("v3/autocomplete?id=TestData");
             var content = await response.Content.ReadAsStreamAsync();
@@ -177,7 +177,7 @@ namespace BaGet.Tests
         [Fact]
         public async Task VersionListReturnsOk()
         {
-            await _factory.AddPackageAsync(_packageStream);
+            await _app.AddPackageAsync(_packageStream);
 
             var response = await _client.GetAsync("v3/package/TestData/index.json");
             var content = await response.Content.ReadAsStringAsync();
@@ -197,7 +197,7 @@ namespace BaGet.Tests
         [Fact]
         public async Task PackageDownloadReturnsOk()
         {
-            await _factory.AddPackageAsync(_packageStream);
+            await _app.AddPackageAsync(_packageStream);
 
             using var response = await _client.GetAsync("v3/package/TestData/1.2.3/TestData.1.2.3.nupkg");
 
@@ -216,7 +216,7 @@ namespace BaGet.Tests
         [Fact]
         public async Task NuspecDownloadReturnsOk()
         {
-            await _factory.AddPackageAsync(_packageStream);
+            await _app.AddPackageAsync(_packageStream);
 
             using var response = await _client.GetAsync(
                 "v3/package/TestData/1.2.3/TestData.1.2.3.nuspec");
@@ -236,7 +236,7 @@ namespace BaGet.Tests
         [Fact]
         public async Task PackageMetadataReturnsOk()
         {
-            await _factory.AddPackageAsync(_packageStream);
+            await _app.AddPackageAsync(_packageStream);
 
             using var response = await _client.GetAsync("v3/registration/TestData/index.json");
             var content = await response.Content.ReadAsStreamAsync();
@@ -311,7 +311,7 @@ namespace BaGet.Tests
         [Fact]
         public async Task PackageMetadataLeafReturnsOk()
         {
-            await _factory.AddPackageAsync(_packageStream);
+            await _app.AddPackageAsync(_packageStream);
 
             using var response = await _client.GetAsync("v3/registration/TestData/1.2.3.json");
             var content = await response.Content.ReadAsStreamAsync();
@@ -365,8 +365,8 @@ namespace BaGet.Tests
         [Fact]
         public async Task SymbolDownloadReturnsOk()
         {
-            await _factory.AddPackageAsync(_packageStream);
-            await _factory.AddSymbolPackageAsync(_symbolPackageStream);
+            await _app.AddPackageAsync(_packageStream);
+            await _app.AddSymbolPackageAsync(_symbolPackageStream);
 
             using var response = await _client.GetAsync(
                 "api/download/symbols/testdata.pdb/16F71ED8DD574AA2AD4A22D29E9C981Bffffffff/testdata.pdb");
@@ -380,8 +380,8 @@ namespace BaGet.Tests
         [InlineData("api/download/symbols/testprefix/testdata.pdb/16F71ED8DD574AA2AD4A22D29E9C981Bffffffff/testdata.pdb")]
         public async Task MalformedSymbolDownloadReturnsOk(string uri)
         {
-            await _factory.AddPackageAsync(_packageStream);
-            await _factory.AddSymbolPackageAsync(_symbolPackageStream);
+            await _app.AddPackageAsync(_packageStream);
+            await _app.AddSymbolPackageAsync(_symbolPackageStream);
 
             using var response = await _client.GetAsync(uri);
 
@@ -395,6 +395,12 @@ namespace BaGet.Tests
                 "api/download/symbols/doesnotexist.pdb/16F71ED8DD574AA2AD4A22D29E9C981Bffffffff/doesnotexist.pdb");
 
             Assert.Equal(HttpStatusCode.NotFound, response.StatusCode);
+        }
+
+        public void Dispose()
+        {
+            _app.Dispose();
+            _client.Dispose();
         }
 
         private string PrettifyJson(Stream jsonStream)
