+++ tests/BaGet.Tests/NuGetClientIntegrationTests.cs
@@ -4,7 +4,6 @@ using System.IO;
 using System.Net.Http;
 using System.Threading;
 using System.Threading.Tasks;
-using Microsoft.AspNetCore.Mvc.Testing;
 using NuGet.Configuration;
 using NuGet.Protocol;
 using NuGet.Protocol.Core.Types;
@@ -17,9 +16,9 @@ namespace BaGet.Tests
     /// <summary>
     /// Uses the official NuGet client to interact with the BaGet test host.
     /// </summary>
-    public class NuGetClientIntegrationTests : IClassFixture<BaGetWebApplicationFactory>
+    public class NuGetClientIntegrationTests : IDisposable
     {
-        private readonly WebApplicationFactory<Startup> _factory;
+        private readonly BaGetApplication _app;
         private readonly HttpClient _client;
 
         private readonly Stream _packageStream;
@@ -29,16 +28,13 @@ namespace BaGet.Tests
         private readonly NuGet.Common.ILogger _logger;
         private readonly CancellationToken _cancellationToken;
 
-        public NuGetClientIntegrationTests(
-            BaGetWebApplicationFactory factory,
-            ITestOutputHelper output)
+        public NuGetClientIntegrationTests(ITestOutputHelper output)
         {
-            _factory = factory.WithOutput(output);
-            _client = _factory.CreateDefaultClient();
-
+            _app = new BaGetApplication(output);
+            _client = _app.CreateDefaultClient();
             _packageStream = TestResources.GetResourceStream(TestResources.Package);
 
-            var sourceUri = new Uri(_factory.Server.BaseAddress, "v3/index.json");
+            var sourceUri = new Uri(_app.Server.BaseAddress, "v3/index.json");
             var packageSource = new PackageSource(sourceUri.AbsoluteUri);
             var providers = new List<Lazy<INuGetResourceProvider>>();
 
@@ -46,7 +42,7 @@ namespace BaGet.Tests
             providers.AddRange(Repository.Provider.GetCoreV3());
 
             _repository = new SourceRepository(packageSource, providers);
-            _cache = new SourceCacheContext() { NoCache = true, MaxAge = new DateTimeOffset(), DirectDownload = true };
+            _cache = new SourceCacheContext { NoCache = true, MaxAge = new DateTimeOffset(), DirectDownload = true };
             _logger = NuGet.Common.NullLogger.Instance;
             _cancellationToken = CancellationToken.None;
         }
@@ -69,7 +65,7 @@ namespace BaGet.Tests
         [Fact]
         public async Task SearchReturnsResults()
         {
-            await _factory.AddPackageAsync(_packageStream);
+            await _app.AddPackageAsync(_packageStream);
 
             var resource = await _repository.GetResourceAsync<PackageSearchResource>();
             var searchFilter = new SearchFilter(includePrerelease: true);
@@ -117,7 +113,7 @@ namespace BaGet.Tests
         [Fact]
         public async Task AutocompleteReturnsResults()
         {
-            await _factory.AddPackageAsync(_packageStream);
+            await _app.AddPackageAsync(_packageStream);
 
             var resource = await _repository.GetResourceAsync<AutoCompleteResource>();
             var results = await resource.IdStartsWith(
@@ -147,7 +143,7 @@ namespace BaGet.Tests
         [Fact]
         public async Task VersionListReturnsResults()
         {
-            await _factory.AddPackageAsync(_packageStream);
+            await _app.AddPackageAsync(_packageStream);
 
             var resource = await _repository.GetResourceAsync<FindPackageByIdResource>();
             var versions = await resource.GetAllVersionsAsync(
@@ -180,7 +176,7 @@ namespace BaGet.Tests
         [InlineData("PackageDoesNotExists", "1.0.0", false)]
         public async Task PackageExistsWorks(string packageId, string packageVersion, bool exists)
         {
-            await _factory.AddPackageAsync(_packageStream);
+            await _app.AddPackageAsync(_packageStream);
 
             var version = NuGetVersion.Parse(packageVersion);
             var resource = await _repository.GetResourceAsync<FindPackageByIdResource>();
@@ -200,7 +196,7 @@ namespace BaGet.Tests
         [InlineData("PackageDoesNotExists", "1.0.0", false)]
         public async Task PackageDownloadWorks(string packageId, string packageVersion, bool exists)
         {
-            await _factory.AddPackageAsync(_packageStream);
+            await _app.AddPackageAsync(_packageStream);
 
             using var packageStream = new MemoryStream();
 
@@ -223,7 +219,7 @@ namespace BaGet.Tests
         [Fact]
         public async Task PackageMetadataReturnsOk()
         {
-            await _factory.AddPackageAsync(_packageStream);
+            await _app.AddPackageAsync(_packageStream);
 
             var resource = await _repository.GetResourceAsync<PackageMetadataResource>();
             var packages = await resource.GetMetadataAsync(
@@ -256,6 +252,13 @@ namespace BaGet.Tests
                 _cancellationToken);
 
             Assert.Empty(packages);
+        }
+
+        public void Dispose()
+        {
+            _app.Dispose();
+            _client.Dispose();
+            _cache.Dispose();
         }
     }
 }
